Stop inheritance and componentize WebSocketClient.
